Adding in-browser tests to create guessing curves.
js/tests/guessing_curve.{js,html}
These files compute a guessing curve for passwords using the in-browser
implementation so that: we know that the browser is actually computing the same
guess numbers and we can see the effect of quantization and compression on the
browser side.

js/src/worker.js
Minor changes about coding style and file names.

utils/quantize_json.py
Utility script to quantize weights in a JSON format file. This is different from
binary quantization because JSON is a text-based (non-binary) format.

utils/find_guess_number.py
Utility script to find the guess number in a guessing list. This is for the use
case of tokenized guessing which can output duplicates due to ambiguous
representations.

git-svn-id: https://cups.cs.cmu.edu/chilisvn/neural-network-cracking/trunk@259 2726a97d-1bfb-4365-8d96-b87bb51a8088